Use builder structs to customize instant/range query parameters
This is mostly an investment in the future to prevent the query and
query_range functions from becoming to bloated in regard to the number
of function parameters.
Future API extensions (e.g. query statistics) will not require a
new major version bump as they can be added in the form of new methods
to the builder structs.