Consider reverse proxies in error handling
If the returned body is JSON, assume its the response from the
Prometheus API and try to decode it, ignoring the status code as the
JSON body may very well contain an API error, e.g. "bad syntax".
If it is not JSON, convert the response to an error and return
it as a wrapped reqwest::Error (this assumes the proxy definitely returns 4xx or 5xx in this case).